Leaderboard-backend/src/main/java/com/group8/Leaderboardbackend/service/LeaderboardRepositoryService.java
package com.group8.Leaderboardbackend.service;

import com.group8.Leaderboardbackend.controller.response.ProfileDto;
import com.group8.Leaderboardbackend.model.Profile;
import com.group8.Leaderboardbackend.repository.ProfileRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;

import java.util.List;
import java.util.Optional;

@Service
@RequiredArgsConstructor
public class LeaderboardRepositoryService {
    private final ProfileRepository profileRepository;

    public List<Profile> getProfiles() {
        return profileRepository.findAll();
    }

    public Profile addUser(Profile profile){
        return profileRepository.save(profile);
    }

}
